LPS-137761 Make window available methods created from Editor component
Instead of defining these methods in each component e.g ClassicEditor/BalloonEditor, the base Editor component will take care of this
