+++ modules/apps/frontend-editor/frontend-editor-ckeditor-web/src/main/resources/META-INF/resources/editor/ClassicEditor.js
@@ -13,153 +13,80 @@
  */
 
 import PropTypes from 'prop-types';
-import React, {useCallback, useEffect, useRef} from 'react';
+import React from 'react';
 
 import {Editor} from './Editor';
 
-const ClassicEditor = React.forwardRef(
-	(
-		{
-			contents = '',
-			editorConfig,
-			initialToolbarSet = 'simple',
-			name,
-			onChange,
-			onChangeMethodName,
-			title,
-			...otherProps
-		},
-		ref
-	) => {
-		const editorRef = useRef();
+const ClassicEditor = ({
+	contents,
+	editorConfig,
+	initialToolbarSet = 'simple',
+	name,
+	title,
+	...otherProps
+}) => {
+	return (
+		<div id={`${name}Container`}>
+			{title && (
+				<label className="control-label" htmlFor={name}>
+					{title}
+				</label>
+			)}
 
-		const getHTML = useCallback(() => {
-			let data = contents;
+			<Editor
+				className="lfr-editable"
+				config={{
+					toolbar: initialToolbarSet,
+					...editorConfig,
+				}}
+				contents={contents}
+				name={name}
+				onBeforeLoad={(CKEDITOR) => {
+					CKEDITOR.disableAutoInline = true;
+					CKEDITOR.dtd.$removeEmpty.i = 0;
+					CKEDITOR.dtd.$removeEmpty.span = 0;
 
-			const editor = editorRef.current.editor;
+					CKEDITOR.getNextZIndex = function () {
+						return CKEDITOR.dialog._.currentZIndex
+							? CKEDITOR.dialog._.currentZIndex + 10
+							: Liferay.zIndex.WINDOW + 10;
+					};
+				}}
+				onDrop={(event) => {
+					const data = event.data.dataTransfer.getData('text/html');
+					const editor = event.editor;
 
-			if (editor && editor.instanceReady) {
-				data = editor.getData();
-
-				if (
-					CKEDITOR.env.gecko &&
-					CKEDITOR.tools.trim(data) === '<br />'
-				) {
-					data = '';
-				}
-
-				data = data.replace(/(\u200B){7}/, '');
-			}
-
-			return data;
-		}, [contents]);
-
-		const onChangeCallback = () => {
-			if (!onChangeMethodName && !onChange) {
-				return;
-			}
-
-			const editor = editorRef.current.editor;
-
-			if (editor.checkDirty()) {
-				if (onChangeMethodName) {
-					window[onChangeMethodName](getHTML());
-				}
-				else {
-					onChange(getHTML());
-				}
-
-				editor.resetDirty();
-			}
-		};
-
-		const editorRefsCallback = useCallback(
-			(element) => {
-				if (ref) {
-					ref.current = element;
-				}
-				editorRef.current = element;
-			},
-			[ref, editorRef]
-		);
-
-		useEffect(() => {
-			window[name] = {
-				getHTML,
-				getText() {
-					return contents;
-				},
-			};
-		}, [contents, getHTML, name]);
-
-		return (
-			<div id={`${name}Container`}>
-				{title && (
-					<label className="control-label" htmlFor={name}>
-						{title}
-					</label>
-				)}
-
-				<Editor
-					className="lfr-editable"
-					config={{
-						toolbar: initialToolbarSet,
-						...editorConfig,
-					}}
-					name={name}
-					onBeforeLoad={(CKEDITOR) => {
-						CKEDITOR.disableAutoInline = true;
-						CKEDITOR.dtd.$removeEmpty.i = 0;
-						CKEDITOR.dtd.$removeEmpty.span = 0;
-
-						CKEDITOR.getNextZIndex = function () {
-							return CKEDITOR.dialog._.currentZIndex
-								? CKEDITOR.dialog._.currentZIndex + 10
-								: Liferay.zIndex.WINDOW + 10;
-						};
-					}}
-					onChange={onChangeCallback}
-					onDrop={(event) => {
-						const data = event.data.dataTransfer.getData(
-							'text/html'
+					if (data) {
+						const fragment = CKEDITOR.htmlParser.fragment.fromHtml(
+							data
 						);
-						const editor = event.editor;
 
-						if (data) {
-							const fragment = CKEDITOR.htmlParser.fragment.fromHtml(
-								data
-							);
+						const name = fragment.children[0].name;
 
-							const name = fragment.children[0].name;
-
-							if (name) {
-								return editor.pasteFilter.check(name);
-							}
+						if (name) {
+							return editor.pasteFilter.check(name);
 						}
-					}}
-					onInstanceReady={({editor}) => {
-						editor.setData(contents, {
-							callback: () => {
-								editor.resetUndo();
-							},
-							noSnapshot: true,
-						});
-					}}
-					ref={editorRefsCallback}
-					{...otherProps}
-				/>
-			</div>
-		);
-	}
-);
+					}
+				}}
+				onInstanceReady={({editor}) => {
+					editor.setData(contents, {
+						callback: () => {
+							editor.resetUndo();
+						},
+						noSnapshot: true,
+					});
+				}}
+				{...otherProps}
+			/>
+		</div>
+	);
+};
 
 ClassicEditor.propTypes = {
 	contents: PropTypes.string,
 	editorConfig: PropTypes.object,
 	initialToolbarSet: PropTypes.string,
 	name: PropTypes.string,
-	onChange: PropTypes.func,
-	onChangeMethodName: PropTypes.string,
 	title: PropTypes.string,
 };
